Extract methods of ReasonStringBuilder in interface
